feat: added hook to change navbar color
